% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chains-package.R
\docType{package}
\name{chains-package}
\alias{chains}
\alias{chains-package}
\title{chains: Complex Reactive Values for 'shiny'}
\description{
Create hybrid reactive objects that can be set imperatively but are validated reactively, ensuring their state is always consistent with their dependencies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://chains.shinyworks.org}
  \item \url{https://github.com/shinyworks/chains}
  \item Report bugs at \url{https://github.com/shinyworks/chains/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

}
\keyword{internal}
