% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_orchestrator.R
\name{choose_orchestrator}
\alias{choose_orchestrator}
\title{Choose an orchestrator}
\usage{
choose_orchestrator(orchestrator, call = rlang::caller_env())
}
\arguments{
\item{orchestrator}{(\code{BatchOrchestrator} or \code{batch_reactive_val}) A
\code{BatchOrchestrator} object, or a \code{batch_reactive_val} that should be
included in the same batch.}

\item{call}{(\code{environment}) The execution environment to use for error
messages.}
}
\value{
A \code{BatchOrchestrator} object.
}
\description{
Choose an orchestrator
}
\keyword{internal}
