% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_to_dt.R
\name{conf_to_dt}
\alias{conf_to_dt}
\title{Convert a list of conf into two data.tables of global and individual features.}
\usage{
conf_to_dt(
  confs,
  allowed_run_info_cols = c("date_init", "date_start_run", "date_end_run", "priority",
    "status"),
  allow_descr = TRUE,
  allowed_function_cols = c("path", "name"),
  allow_args = TRUE
)
}
\arguments{
\item{confs}{\code{list of list}. List of conf list(s) from yaml file(s).}

\item{allowed_run_info_cols}{\code{character} (c("date_init", "date_start_run", "date_end_run", "priority", "status")). Run info elements to be kept.}

\item{allow_descr}{\code{boolean or character} (TRUE). Either a boolean specifying whether or not to keep descr elements, or column names.}

\item{allowed_function_cols}{\code{character} (c("names", "path")). Function elements to be kept.}

\item{allow_args}{\code{boolean or character} (TRUE). Either a boolean specifying whether or not to keep args elements, or column names.}
}
\value{
a list of two tables 'tbl_global' and 'tbl_idv':
\itemize{
 \item{tbl_global} {contains the global features of the configurations (.$run_info and .$descriptive)}
 \item{tbl_idv} {contains the individual features of the configurations (.$function and .$args)}
}
}
\description{
Convert a list of conf into two data.tables of global and individual features.
}
\examples{
\dontrun{\donttest{

conf_1 <- list(
  "run_info" = list(
    "date_init" = as.character(Sys.time()),
    "date_start_run" = "N/A",
    "date_end_run" = "N/A",
    "priority" = 1,
    "status" = "waiting"),
  "descriptive" = list("descr_1" = "d1.1",
                       "descr_2" = "d1.2"),
  "function" = list(
    "path" = "fun_path_1",
    "name" = "fun_name_1"
  ),
  "args" = list(
    x = "1",
    y = "TRUE"
  )
)

conf_2 <- list(
  "run_info" = list(
    "date_init" = as.character(Sys.time() + 1),
    "date_start_run" = "N/A",
    "date_end_run" = "N/A",
    "priority" = 2,
    "status" = "waiting"),
  "descriptive" = list("descr_1" = "d2.1",
                       "descr_3" = "d2.3"),
  "function" = list(
    "path" = "fun_path_2",
    "name" = "fun_name_2"
  ),
  "args" = list(
    x = 2,
    y = "/path/conf_2/inputs/y.RDS"
  )
)

confs <- list(conf_1, conf_2)

conf_to_dt(confs)
conf_to_dt(confs, 
           allow_descr = F,
           allow_args = F)
conf_to_dt(confs,
           allowed_run_info_cols = c("date_init", "status"),
           allowed_function_cols = c("path"))
conf_to_dt(confs,
           allowed_run_info_cols = NULL,
           allowed_function_cols = NULL)
conf_to_dt(confs,
           allowed_run_info_cols = "",
           allowed_function_cols = "",
           allow_descr = F,
           allow_args = F)

}} 
}
