% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validated_reactive_val.R
\name{.catch_vrv_error}
\alias{.catch_vrv_error}
\title{Capture a vrv error without re-throwing it}
\usage{
.catch_vrv_error(expr)
}
\arguments{
\item{expr}{The expression to evaluate.}
}
\value{
The result of the expression, or the "defused" condition if an
error is thrown.
}
\description{
Capture a vrv error without re-throwing it
}
\keyword{internal}
