% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validated_reactive_val.R
\name{.create_validation_reactive}
\alias{.create_validation_reactive}
\title{Create the Core Validation Reactive}
\usage{
.create_validation_reactive(validation_expr, state_rv, label, env)
}
\arguments{
\item{validation_expr}{(\code{expression}) A reactive expression that defines the
authoritative state of this \code{validated_reactive_value}, This expression can
access the \code{validated_reactive_value}'s own current value via a \code{self()}
\code{\link[shiny:reactiveVal]{shiny::reactiveVal()}} to reconcile it with upstream dependencies.}

\item{state_rv}{(\code{reactiveVal}) The \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}} that holds the
state.}

\item{label}{(length-1 \code{character} or \code{NULL}) An optional label for the
\code{validated_reactive_value}, used for debugging.}

\item{env}{(\code{environment}) The environment in which to evaluate the
\code{validation_expr}.}
}
\value{
A \code{\link[shiny:reactive]{shiny::reactive()}}.
}
\description{
Create the core \code{\link[shiny:reactive]{shiny::reactive()}} expression that runs the user-provided
validation logic, and catch any errors introduce by that logic.
}
\keyword{internal}
