% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validated_reactive_val.R
\name{.create_vrv_function}
\alias{.create_vrv_function}
\title{Create the User-Facing Getter/Setter Function}
\usage{
.create_vrv_function(rvs, validation_rctv, default_quo)
}
\arguments{
\item{validation_rctv}{(\code{reactive}) The \code{\link[shiny:reactive]{shiny::reactive()}} that performs the
validation.}
}
\value{
A function that acts as a getter and setter.
}
\description{
Create the user-facing function. Without an argument, the resulting function
triggers the validation reactive. With an argument, the resulting function
imperatively updates the internal state.
}
\keyword{internal}
