% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validated_reactive_val.R
\name{.enquo_validation_expr}
\alias{.enquo_validation_expr}
\title{Prepare the validation quosure}
\usage{
.enquo_validation_expr(validation_expr, value_rv, env)
}
\arguments{
\item{validation_expr}{(\code{expression}) A reactive expression that defines the
authoritative state of this \code{validated_reactive_value}, This expression can
access the \code{validated_reactive_value}'s own current value via the \code{.vrv()}
pronoun to reconcile it with upstream dependencies.}

\item{value_rv}{(\code{reactiveVal}) The \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}} that holds the
current value.}

\item{env}{(\code{environment}) The environment in which to evaluate the
\code{validation_expr}.}
}
\value{
The \code{validation_expr} enclosed in an environment that is a child of
\code{env}, with a \code{.vrv()} pronoun function.
}
\description{
Capture the user-provided validation expression and set its environment to a
new child environment containing the \code{.vrv} pronoun.
}
\keyword{internal}
