% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_value_maybe_reactive}
\alias{.get_value_maybe_reactive}
\title{Get a value that might be reactive}
\usage{
.get_value_maybe_reactive(x)
}
\arguments{
\item{x}{The value, which might be a reactive expression.}
}
\value{
The value, or the result of evaluating the reactive expression.
}
\description{
Get a value that might be reactive
}
\keyword{internal}
