% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validated_reactive_val.R
\name{.handle_validation_result}
\alias{.handle_validation_result}
\title{Handle the result of a validation expression}
\usage{
.handle_validation_result(result, default_quo)
}
\arguments{
\item{result}{The value returned from the validation reactive.}

\item{default_quo}{The quosure for the default value.}
}
\value{
A list with elements \code{value}, \code{error}, and \code{is_default}.
}
\description{
Check if the validation result is an error, and return a list containing the
final value and its status.
}
\keyword{internal}
