% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validated_reactive_val.R
\name{.handle_validation_success}
\alias{.handle_validation_success}
\title{Handle a validation success}
\usage{
.handle_validation_success(result, rvs)
}
\arguments{
\item{result}{The validated value.}

\item{rvs}{A list of the internal \code{reactiveVal}s.}
}
\value{
The validated value.
}
\description{
Sets the internal state \code{reactiveVal}s to reflect a validation success.
}
\keyword{internal}
