% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validated_reactive_val.R
\name{.initialize_rvs}
\alias{.initialize_rvs}
\title{Create and Label the Internal \code{reactiveVal}s}
\usage{
.initialize_rvs(value = NULL, label = NULL)
}
\arguments{
\item{value}{(various) The initial value.}

\item{label}{(length-1 \code{character} or \code{NULL}) An optional label for the
\code{validated_reactive_value}, used for debugging.}
}
\value{
A list of \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}}s.
}
\description{
Create the internal \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}}s that hold the state for a
\code{validated_reactive_val()}.
}
\keyword{internal}
