% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validated_reactive_val.R
\name{.initialize_value_rv}
\alias{.initialize_value_rv}
\title{Create and Label the Internal \code{reactiveVal}}
\usage{
.initialize_value_rv(value = NULL, label = NULL)
}
\arguments{
\item{value}{(various) The initial value. This value will be coerced via the
validation expression when accessed. If this value is reactive, an observer
with \code{priority = Inf} will be created to attempt to keep the validated
value in sync with that reactive.}

\item{label}{(length-1 \code{character} or \code{NULL}) An optional label for the
\code{validated_reactive_value}, used for debugging.}
}
\value{
A \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}}.
}
\description{
Create the internal \code{\link[shiny:reactiveVal]{shiny::reactiveVal()}} that holds the state for a
\code{validated_reactive_val()}.
}
\keyword{internal}
