% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validated_reactive_val.R
\name{.sync_vrv_to_value}
\alias{.sync_vrv_to_value}
\title{Update vrv when value changes}
\usage{
.sync_vrv_to_value(vrv_fun, value)
}
\arguments{
\item{vrv_fun}{(\code{vrv}, \code{function}) A function to get and set validated reactive
values.}

\item{value}{(various) The initial value. This value will be coerced via the
validation expression when accessed. If this value is reactive, an observer
with \code{priority = Inf} will be created to attempt to keep the validated
value in sync with that reactive.}
}
\value{
A \code{\link[shiny:observe]{shiny::observe()}} with \code{priority = Inf} to reactively call
\code{vrv_fun()} with the value of \code{value} whenever \code{value} changes.
}
\description{
Update vrv when value changes
}
\keyword{internal}
