% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pronoun.R
\name{.vrv}
\alias{.vrv}
\title{Access the validated reactive value pronoun}
\usage{
.vrv(env = rlang::caller_env())
}
\arguments{
\item{env}{(\code{environment}) The environment in which to find the \code{.vrv}
pronoun. This should generally not be changed from the default of
\code{\link[rlang:stack]{rlang::caller_env()}}.}
}
\value{
The current value of the \code{\link[=validated_reactive_val]{validated_reactive_val()}}.
}
\description{
When used inside the \code{validation_expr} of a \code{\link[=validated_reactive_val]{validated_reactive_val()}}, this
function acts as a pronoun to access the current value (before validation) of
that \code{validated_reactive_val}. This allows the validation expression to
reconcile the object's current state with its reactive dependencies.
}
\details{
To reference \code{.vrv()} in a package and avoid \verb{R CMD check} notes, you can
either import this function with \verb{#' @importFrom shinybatch .vrv} or call it
with the \code{shinybatch::.vrv()} namespace.
}
