% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_conf.R
\name{init_task_conf}
\alias{init_task_conf}
\title{Initialise conf file for batch tasks}
\usage{
init_task_conf(
  dir_path,
  conf_descr = NULL,
  fun_path,
  fun_name,
  fun_args = NULL,
  priority = 0L,
  compress = TRUE
)
}
\arguments{
\item{dir_path}{\code{character}. Where to create the new directory.}

\item{conf_descr}{\code{named list} (NULL). Description fields chosen by the user.}

\item{fun_path}{\code{character}. Path to the script of the function.}

\item{fun_name}{\code{character}. Name of the function in fun_path script.}

\item{fun_args}{\code{named list} (NULL). Args of the function, must all be named.}

\item{priority}{\code{numeric} (0L). Number used to define which task should be launched first.}

\item{compress}{For \code{saveRDS}. a logical specifying whether saving to a named file is to use "gzip" compression, or one of "gzip", "bzip2" or "xz" to indicate the type of compression to be used. Ignored if file is a connection.}
}
\value{
a list containing the conf fields. Attribute 'path' of the result contains the path to the conf directory.
}
\description{
Initialise conf file for batch tasks
}
\examples{
\donttest{\dontrun{

# create temporary directory
dir <- tempdir()

# create and save conf
conf <- init_task_conf(dir_path = dir,
                  conf_descr = list(title = "my_title",
                                    description = "my_descr"),
                  fun_path = "my_fun_path",
                  fun_name = "my_fun_name",
                  fun_args = list(x = 1,
                                  y = 0:4,
                                  z = iris),
                  priority = 1)

# catch results
list.files(attr(conf, "path"))
read_conf <- yaml::read_yaml(paste0(attr(conf, "path"), "/", "conf.yml"))
y <- readRDS(paste0(attr(conf, "path"), "inputs/y.RDS"))
z <- readRDS(paste0(attr(conf, "path"), "inputs/z.RDS"))

}}
}
