% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{input_btns}
\alias{input_btns}
\title{create button for DT table}
\usage{
input_btns(
  inputId,
  col_value,
  tooltip,
  label = "",
  icon = NULL,
  status = "primary"
)
}
\arguments{
\item{inputId}{\code{character}.\code{\link{actionButton}}.}

\item{col_value}{\code{vector}. \code{\link{actionButton}}.}

\item{tooltip}{\code{character}. \code{\link{actionButton}}.}

\item{label}{\code{character}. \code{\link{actionButton}}.}

\item{icon}{\code{character}. \code{\link{actionButton}}.}

\item{status}{\code{character}. \code{\link{actionButton}}.}
}
\value{
a shiny btn.
}
\description{
create button for DT table
}
