% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_configure_task.R
\name{configure_task_server}
\alias{configure_task_server}
\title{Module shiny to configure a task.}
\usage{
configure_task_server(
  input,
  output,
  session,
  btn,
  dir_path,
  fun_path,
  fun_name,
  conf_descr = NULL,
  fun_args = NULL,
  priority = 0L,
  compress = TRUE,
  labels = list(success = "Task configured !", error = "Error when configuring the task")
)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny input}

\item{session}{shiny input}

\item{btn}{\code{reactive}. Link to a actionButton for call \code{configure_task} function}

\item{dir_path}{\code{character}. Tasks location (parent directory).}

\item{fun_path}{\code{character}. Path to the script of the function.}

\item{fun_name}{\code{character}. Name of the function in fun_path script.}

\item{conf_descr}{\code{named list} (NULL). Optional description fields.}

\item{fun_args}{\code{named list} (NULL). Args of the function, must all be named.}

\item{priority}{\code{numeric} (0L). Number used to define which task should be launched first.}

\item{compress}{\code{logical or character} (TRUE). Either a logical specifying whether or not to use "gzip" compression, or one of "gzip", "bzip2" or "xz" to indicate the type of compression to be used.}

\item{labels}{\code{list}. Modal dialog title.}
}
\value{
Nothing.
}
\description{
Module shiny to configure a task.
}
\examples{

\donttest{

if(interactive()){

require(shiny)

# create temporary directory for conf
dir_conf <- paste0(tempdir(), "/conf", round(runif(n = 1, max = 10000)))
dir.create(dir_conf, recursive = TRUE)

# ex fun
fun_path = system.file("ex_fun/sb_fun_ex.R", package = "shinybatch")
fun_name = "sb_fun_ex"

# create and save conf
ui <- shiny::fluidPage(
  fluidRow(
    column(offset = 3, 6,
           shiny::actionButton("conf_task", "Configure the task", width = "100\%")
    )
  )
)
server <- function(input, output, session) {
  callModule(configure_task_server, "my_id_1",
             btn = reactive(input$conf_task),
             dir_path = dir_conf,
             conf_descr = list(title = "my_title",
                               description = "my_descr"),
             fun_path = fun_path,
             fun_name = fun_name,
             fun_args = list(x = 1,
                             y = 0:4,
                             z = iris),
             priority = 1)
}
shiny::shinyApp(ui = ui, server = server)

# catch results
list.files(path <- list.dirs(dir_conf, full.names = TRUE, recursive = FALSE))
path
read_conf <- yaml::read_yaml(paste0(path[1], "/", "conf.yml"))
y <- readRDS(paste0(path[1], "/", "inputs/y.RDS"));y
z <- readRDS(paste0(path[1], "/", "inputs/z.RDS"));z

}
}

}
\seealso{
\code{\link[shinybatch]{tasks_overview_server}}
}
