% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_demo_app.R
\name{run_demo_app}
\alias{run_demo_app}
\title{Launch shinybatch demo app}
\usage{
run_demo_app()
}
\description{
Launch shinybatch demo app
}
\details{
The created sheduler is automatically destroyed when quitting Shiny.
}
