% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinybatch-package.R
\docType{package}
\name{shinybatch-package}
\alias{shinybatch}
\alias{shinybatch-package}
\title{shinybatch: Create Validated Reactive Values in 'shiny'}
\description{
Create a hybrid reactive object that can be set imperatively but is validated reactively, ensuring its state is always consistent with its dependencies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://shinyworks.github.io/shinybatch/}
  \item \url{https://github.com/shinyworks/shinybatch}
  \item Report bugs at \url{https://github.com/shinyworks/shinybatch/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

}
\keyword{internal}
