% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrv_character.R
\name{vrv_character}
\alias{vrv_character}
\alias{vrv_character_scalar}
\title{Create a validated reactive character vector}
\usage{
vrv_character(
  value = NULL,
  default = NULL,
  allow_null = TRUE,
  allow_na = TRUE,
  min_size = NULL,
  max_size = NULL,
  regex = NULL,
  label = NULL,
  env = rlang::caller_env()
)

vrv_character_scalar(
  value = NULL,
  default = NULL,
  label = NULL,
  allow_null = TRUE,
  allow_zero_length = TRUE,
  allow_na = TRUE,
  regex = NULL,
  env = rlang::caller_env()
)
}
\arguments{
\item{value}{(various) The initial value.}

\item{default}{(\code{expression}) A value to use when the current value is not
valid according to the rules. Can be a reactive expression. Defaults to
\code{NULL}.}

\item{allow_null}{\verb{(length-1 logical)} Is NULL an acceptable value?}

\item{allow_na}{\verb{(length-1 logical)} Are NA values ok?}

\item{min_size}{\verb{(length-1 integer)} The minimum size of the object. Object
size will be tested using \code{\link[vctrs:vec_size]{vctrs::vec_size()}}.}

\item{max_size}{\verb{(length-1 integer)} The maximum size of the object. Object
size will be tested using \code{\link[vctrs:vec_size]{vctrs::vec_size()}}.}

\item{regex}{\verb{(character, list, or stringr_pattern)} One or more optional
regular expressions to test against the values of \code{x}. This can be a
character vector, a list of character vectors, or a pattern object from the
\{stringr\} package (e.g., \code{stringr::fixed("a.b")}). The default error
message for non-matching values will include the pattern itself (see
\code{\link[stbl:regex_must_match]{regex_must_match()}}). To provide a custom message, supply a named
character vector where the value is the regex pattern and the name is the
message that should be displayed. To check that a pattern is \emph{not} matched,
attach a \code{negate} attribute set to \code{TRUE}. If a complex regex pattern
throws an error, try installing the stringi package.}

\item{label}{(length-1 \code{character} or \code{NULL}) An optional label for the
\code{validated_reactive_value}, used for debugging.}

\item{env}{(\code{environment}) The environment in which to evaluate the
\code{validation_expr}.}

\item{allow_zero_length}{\verb{(length-1 logical)} Are zero-length vectors
acceptable?}
}
\value{
A \code{vrv} object which returns a validated character vector.
}
\description{
A wrapper around \code{\link[=validated_reactive_val]{validated_reactive_val()}} that uses \code{\link[stbl:stabilize_chr]{stbl::stabilize_chr()}}
to validate and coerce its value. This is a convenience function that
constructs the \code{validation_expr} for you.
}
