% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrv_factor.R
\name{vrv_factor}
\alias{vrv_factor}
\title{Create a validated reactive factor}
\usage{
vrv_factor(
  levels,
  default,
  value = NULL,
  label = NULL,
  env = rlang::caller_env()
)
}
\arguments{
\item{levels}{(\code{expression}) A reactive expression that returns a \code{character}
vector of valid levels.}

\item{default}{(\code{expression}) A reactive expression that returns a
length-1 \code{character} vector to use as the default value when the current
value is not in \code{levels}.}

\item{value}{(various) The initial value.}

\item{label}{(length-1 \code{character} or \code{NULL}) An optional label for the
\code{validated_reactive_value}, used for debugging.}

\item{env}{(\code{environment}) The environment in which to evaluate the
\code{validation_expr}.}
}
\value{
A \code{validated_reactive_val} object.
}
\description{
A wrapper around \code{\link[=validated_reactive_val]{validated_reactive_val()}} to help manage factors,
especially in a cascading / dependent manner. This is a convenience function
that constructs the \code{validation_expr} for you.
}
