% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrv_from_function.R
\name{vrv_from_function}
\alias{vrv_from_function}
\title{Create a validated reactive value from a validation function}
\usage{
vrv_from_function(
  validation_fn,
  ...,
  value = NULL,
  default = NULL,
  label = NULL,
  env = rlang::caller_env()
)
}
\arguments{
\item{validation_fn}{(\code{function}) A function to use for validation.}

\item{...}{Quosures of arguments to pass to \code{validation_fn}.}

\item{value}{(various) The initial value. This value will be coerced via the
validation expression when accessed. If this value is reactive, an observer
with \code{priority = Inf} will be created to attempt to keep the validated
value in sync with that reactive.}

\item{default}{(various, including \code{expression} or \verb{reactive expression}) A
value to use when the current value is not valid according to the defined
rules. Defaults to \code{NULL}.}

\item{label}{(length-1 \code{character} or \code{NULL}) An optional label for the
\code{validated_reactive_value}, used for debugging.}

\item{env}{(\code{environment}) The environment in which to evaluate the
\code{validation_expr}.}
}
\value{
A \code{vrv} object.
}
\description{
A wrapper around \code{\link[=validated_reactive_val]{validated_reactive_val()}} that uses a provided function
to validate and coerce its value. This is a convenience function that
constructs the \code{validation_expr} for you.
}
\keyword{internal}
