% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_task.R
\name{configure_task}
\alias{configure_task}
\title{Initialise conf file for batch tasks}
\usage{
configure_task(dir_path, fun_path, fun_name, conf_descr = NULL,
  fun_args = NULL, priority = 0L, compress = TRUE)
}
\arguments{
\item{dir_path}{\code{character}. Where to create the new directory.}

\item{fun_path}{\code{character}. Path to the script of the function.}

\item{fun_name}{\code{character}. Name of the function in fun_path script.}

\item{conf_descr}{\code{named list} (NULL). Description fields chosen by the user.}

\item{fun_args}{\code{named list} (NULL). Args of the function, must all be named.}

\item{priority}{\code{numeric} (0L). Number used to define which task should be launched first.}

\item{compress}{\code{logical or character} (TRUE). Either a logical specifying whether or not to use "gzip" compression, or one of "gzip", "bzip2" or "xz" to indicate the type of compression to be used.}
}
\value{
a list containing the conf fields. Attribute 'path' of the result contains the path to the conf directory.
}
\description{
Initialise conf file for batch tasks
}
\examples{
\donttest{\dontrun{

# create temporary directory
dir <- tempdir()

# create and save conf
conf <- configure_task(dir_path = dir,
                       conf_descr = list(title = "my_title",
                                         description = "my_descr"),
                       fun_path = "my_fun_path",
                       fun_name = "my_fun_name",
                       fun_args = list(x = 1,
                                       y = 0:4,
                                       z = iris),
                       priority = 1)

# catch results
list.files(conf$dir)
read_conf <- yaml::read_yaml(paste0(conf$dir, "/", "conf.yml"))
y <- readRDS(paste0(conf$dir, "inputs/y.RDS"))
z <- readRDS(paste0(conf$dir, "inputs/z.RDS"))

}}
}
