% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validated_reactive_val.R
\name{.handle_validation_error}
\alias{.handle_validation_error}
\title{Handle a validation error}
\usage{
.handle_validation_error(result, rvs, default_quo)
}
\arguments{
\item{result}{The captured error condition.}

\item{rvs}{A list of the internal \code{reactiveVal}s.}

\item{default_quo}{The quosure for the default value.}
}
\value{
The evaluated default value.
}
\description{
Sets the internal state \code{reactiveVal}s to reflect a validation failure.
}
\keyword{internal}
