% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validated_reactive_val.R
\name{.handle_validation_result}
\alias{.handle_validation_result}
\title{Handle the result of a validation expression}
\usage{
.handle_validation_result(result, rvs, default_quo)
}
\arguments{
\item{result}{The value returned from the validation reactive.}

\item{rvs}{A list of the internal \code{reactiveVal}s.}

\item{default_quo}{The quosure for the default value.}
}
\value{
The validated value or the default value.
}
\description{
Check if the validation result is an error, update the internal state
\code{reactiveVal}s accordingly, and return the final value (either the validated
result or the default).
}
\keyword{internal}
