% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_tasks_overview.R
\name{tasks_overview_server}
\alias{tasks_overview_server}
\alias{tasks_overview_UI}
\title{Module to visualize all the created tasks.}
\usage{
tasks_overview_server(
  input,
  output,
  session,
  dir_path,
  allowed_run_info_cols = c("date_init", "date_start_run", "date_end_run", "priority",
    "status"),
  allow_descr = TRUE,
  allowed_function_cols = c("path", "name"),
  allow_args = TRUE,
  allowed_status = c("waiting", "running", "finished", "error"),
  update_mode = c("reactive", "button"),
  intervalMillis = 1000,
  table_fun = function(x) x,
  ...
)

tasks_overview_UI(
  id,
  labels = list(btn = "Display/update the tasks", empty_global =
    "Empty table of global features.", empty_individual =
    "Empty table of individual features.", error_dir_access =
    "Cannot access given directory.")
)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny input}

\item{session}{shiny input}

\item{dir_path}{\code{character}. Where to find the tasks directorys.}

\item{allowed_run_info_cols}{\code{character} (c("date_init", "date_start_run", "date_end_run", "priority", "status")). Run info elements to be kept.}

\item{allow_descr}{\code{boolean or character} (TRUE). Either a boolean specifying whether or not to keep descr elements, or column names.}

\item{allowed_function_cols}{\code{character} (c("names", "path")). Function elements to be kept.}

\item{allow_args}{\code{boolean or character} (TRUE). Either a boolean specifying whether or not to keep args elements, or column names.}

\item{allowed_status}{\code{character} (c("waiting", "running", "finished", "error")). Vector of allowed status.}

\item{update_mode}{\code{character} : "reactive" (default) use \code{shiny::reactivePoll} to see if tasks info have changed. 
"button" add and use a \code{shiny::actionButton} to update task info table.}

\item{intervalMillis}{\code{integer}. In case of "reactive" update_mode, time betweens calls}

\item{table_fun}{\code{function} (function(x) x). Function to be applied on the summary table, making it easy to customize it. First arg must be the summmary table.}

\item{...}{\code{}. Additional args to be given to the table_fun function.}

\item{id}{\code{character}. shiny id to allow multiple instanciation.}

\item{labels}{\code{list} UI labels}
}
\value{
the status of the selected line (one run) of the summmary table and the path to the directory in which its output is stored.
}
\description{
Module to visualize all the created tasks.
}
\examples{
\dontrun{\donttest{

# create temporary directory for conf
dir_conf <- paste0(tempdir(), "/conf")
dir.create(dir_conf, recursive = T)

# ex fun 
fun_path = system.file("ex_fun/sb_fun_ex.R", package = "shinybatch")
fun_name = "sb_fun_ex"

# create 2 confs
conf_1 <- configure_task(dir_path = dir_conf,
                         conf_descr = list(title = "my_title_1",
                                           description = "my_descr_1"),
                         fun_path = "my_fun_path_1",
                         fun_name = "my_fun_name_1",
                         fun_args = list(x = 1,
                                         y = 0:4,
                                         z = iris),
                         priority = 1)
conf_2 <- configure_task(dir_path = dir_conf,
                         conf_descr = list(title = "my_title_2",
                                           description = "my_descr_2"),
                         fun_path = fun_path,
                         fun_name = fun_name,
                         fun_args = list(x = 1,
                                         y = 0:4,
                                         z = iris),
                         priority = 2)
                         
run_task(paste0(conf_2$dir, "conf.yml"))

ui <- shiny::fluidPage(tasks_overview_UI("my_id_1"))
server <- function(input, output, session) {
  callModule(tasks_overview_server, "my_id_1",
             dir_path = dir_conf,
             allowed_status = c("waiting", "running", "finished", "error"),
             allowed_run_info_cols = NULL,
             allowed_function_cols = "",
             allow_descr = T,
             allow_args = T,
             table_fun = function(x, y) x[, new_col := y],
             y = "created using table_fun")
}
shiny::shinyApp(ui = ui, server = server)

}}

}
