% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_order.R
\name{run_order}
\alias{run_order}
\title{Compute the order of priority to run a list of configurations}
\usage{
run_order(confs)
}
\arguments{
\item{confs}{\code{list}. .}
}
\value{
the order of priority of the given confs according to their priority argument and
date of creation.
}
\description{
Compute the order of priority to run a list of configurations
}
