% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-shared.R
\name{shared-params}
\alias{shared-params}
\title{Parameters used in multiple functions}
\arguments{
\item{call}{(\code{environment}) The execution environment to mention as the
source of error messages.}

\item{message_env}{(\code{environment}) The execution environment to use to
evaluate variables in error messages.}

\item{timeout}{(length-1 \code{integer}) The maximum number of seconds to wait for
a transaction to complete.}
}
\description{
Reused parameter definitions are gathered here for easier editing. R6 methods
can't \code{inheritParams}, so you still need to copy/paste to the individual
methods.
}
\keyword{internal}
