% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinybatch-package.R
\docType{package}
\name{shinybatch-package}
\alias{shinybatch}
\alias{shinybatch-package}
\title{shinybatch: Batch Shiny Reactive Updates}
\description{
A new type of reactive object for tightly coupled variables in 'shiny' apps.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://shinyworks.github.io/shinybatch/}
  \item \url{https://github.com/shinyworks/shinybatch}
  \item Report bugs at \url{https://github.com/shinyworks/shinybatch/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

}
\keyword{internal}
