% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrv_factor.R
\name{vrv_factor}
\alias{vrv_factor}
\alias{vrv_factor_scalar}
\title{Create a validated reactive factor-like character vector}
\usage{
vrv_factor(
  levels,
  value = NULL,
  default = NULL,
  allow_null = TRUE,
  allow_na = TRUE,
  min_size = NULL,
  max_size = NULL,
  to_na = character(),
  label = NULL,
  env = rlang::caller_env()
)

vrv_factor_scalar(
  levels,
  value = NULL,
  default = NULL,
  allow_null = TRUE,
  allow_zero_length = TRUE,
  allow_na = TRUE,
  to_na = character(),
  label = NULL,
  env = rlang::caller_env()
)
}
\arguments{
\item{levels}{(\code{expression}) An expression that returns a \code{character} vector
of valid levels. Can be a reactive expression.}

\item{value}{(various) The initial value. This value will be coerced via the
validation expression when accessed.}

\item{default}{(various, including \code{expression} or \verb{reactive expression}) A
value to use when the current value is not valid according to the defined
rules. Defaults to \code{NULL}.}

\item{allow_null}{(length-1 \code{logical}, \code{expression}, or \verb{reactive expression}) Is \code{NULL} an acceptable value?}

\item{allow_na}{(length-1 \code{logical}, \code{expression}, or \verb{reactive expression})
Are \code{NA} values okay?}

\item{min_size}{(length-1 \code{integer}, \code{expression}, or \verb{reactive expression})
The minimum size of the vector.}

\item{max_size}{(length-1 \code{integer}, \code{expression}, or \verb{reactive expression})
The maximum size of the vector.}

\item{to_na}{(\code{expression}) Values to coerce to \code{NA}. Can be a reactive
expression.}

\item{label}{(length-1 \code{character} or \code{NULL}) An optional label for the
\code{validated_reactive_value}, used for debugging.}

\item{env}{(\code{environment}) The environment in which to evaluate the
\code{validation_expr}.}

\item{allow_zero_length}{(\code{expression}) Is a zero-length vector acceptable?
Can be a reactive expression.}
}
\value{
A \code{vrv} object which returns a factor-like \code{character} vector.
}
\description{
A wrapper around \code{\link[=validated_reactive_val]{validated_reactive_val()}} to help manage factor-like
values, especially in a cascading / dependent manner. This is a convenience
function that constructs the \code{validation_expr} for you by wrapping
\code{\link[stbl:stabilize_fct]{stbl::stabilize_fct()}}. Note that the objects returned by the resulting
\code{vrv} are \code{character} vectors, not \code{factor}s, to allow values to remain
unchanged when the allowed levels change but the value is still valid.
}
